basic user-role implementation

+++ backend/src/main/java/br/ucs/greencitizenship/projections/UserDetailsProjection.java
@@ -3,6 +3,6 @@ package br.ucs.greencitizenship.projections;
 public interface UserDetailsProjection {
     String getUsername();
     String getPassword();
-    Long getRoleId();
+    Integer getRoleId();
     String getAuthority();
 }
